+++ crates/wdl-engine/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Fixed
+
+* Fixed permissions issue in the Docker backend when a container runs with a
+  different user ([#605](https://github.com/stjude-rust-labs/wdl/pull/605)).
+
 ## 0.8.0 - 09-15-2025
 
 #### Added